power: qpnp-smbcharger: enable AICL while unpreparing
Currently, the driver ends up writing a 0 to the enable bit when it
wants to enable AICL. This causes AICL to remain disabled and could
cause the input to collapse. Driver should write a 1 to that enable
bit when it wants to enable AICL.

Also it is safe to enable the AICL bit before faking an insertion,
so move the enabling of AICL before that.

CRs-Fixed: 991066
Change-Id: I3d5730c5ab81da142bb110bce93e0ba71c6c3e9a
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Vamshi Krishna B V <vbv@codeaurora.org>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>