scsi: ufs-debugfs: change the root dir name to device name
Currently we are using the "ufshcdXY" (where XY would be a number starting
from 0) format for debugfs root directory for each host controller
instances. But this numbering scheme isn't very informative and now that
we will be having 2 UFS host controllers in future targets (one for
embedded UFS and second one for removable UFS card), it would be good
to change the root directory name as per the UFS host controller device
name. This name would be like "<base-address>.ufshc_mem" or
"<base-address>.ufshc_card" which would be more readable.

Change-Id: Iadaeec282e78128278539de561a0d67ee0b7d7e7
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>