scsi: ufs: prefix host device name for hex dumps
ufshcd_hex_dump() is the function used to print out the hex dumps for
debugging purpose and currently we aren't printing which host controller
instance these hex dumps relate to. Now that we might have SOCs having
multiple UFS host controllers, it is important to print out the host
controller device name as well to know the context of these hex dumps.

Change-Id: I238512afdc2734e99e5821da76af1aea6d22c456
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>