arm/dts: msm8996: Mark IRQ 168 as summary IRQ
Some IRQs are disabled lazily, i.e., the IRQ is not disabled until the
next time that an interrupt is raised. This means that some summary IRQs
will fire twice and cause a wakeup that may appear many times and is
listed in a blind metrics query. This changes the IRQ messaging for IRQ
id 168 to reflect that it's a summary IRQ for easier power debugging.

Bug: 34961013
Change-Id: I10478c9cd58181dd7144ac53ddf8533b24f75e77
Signed-off-by: Maggie White <maggiewhite@google.com>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>