ARM: dts: msm: Rework and unify msm-thermal setup
For unknown reasons Qualcomm decided to use very
different msm-thermal configurations on msm8996
and msm8996pro.
The msm8996pro setup would run polling more often
and throttle later than the msm8996 setup would.

As solution, create an improved setup which can
be used by msm8996 and msm8996pro.
For this setup we stick with the recommended polling
rate of 100ms.
We use the core thresholds from msm8996pro.
These are used only when thermal-engine
is unavailable.
The first thermal trigger threshold (qcom,limit-temp)
gets increased by 10degC as the value of 60degC was
too conservative.

Also as additional improvement we decrease the hysteresis
for freq mitigation to 30degC to introduce a timeframe
(between 40degC and 30degC) where the cores get replugged
but the frequency is still mitigated.

Change-Id: Ib345ef6bc5cbdb806589dd6917162fc84fce3566
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>