cpufreq: times: concurrent_active_time by pid
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each pid spent running
with 1 to num_possible_cpus online and not idle.

We can use this information to blame pids and
uids for the energy they are using.

For instance, if 4 cores are online and active
for 50ms and the pid is running on one of the
cores, the pid should be blamed for 1/4 of the
base energy used by the cpu when 1 or more cores is
active. Using other proc files, the pid can
also be blamed for the cost of keeping a cluster
on and the cost of the individual cores it
is using.

This patch is based on commit
a88a869505d33bb77ee7bca67602f5fe6048205d in the
android-msm-wahoo-4.4 branch by Marissa Wall.

It differs from the previous (reverted) commit
b841887ffb80 on android-msm-bluecross-4.9 in that
initialization & allocation of p->concurrent_active_time
is split into separate functions.

Bug: 68399339
Bug: 74242328
Test: cat /proc/<PID>/concurrent_active_time
Change-Id: I3ee466278e0c8742777bed7f3fc5e58648920550
Signed-off-by: Connor O'Brien <connoro@google.com>
Signed-off-by: Davide Garberi <dade.garberi@gmail.com>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>