scsi: ufs: delay switching off regulators
Don't turn-off regulators immediately in case of errors.
If eMMC is not yet probed, turing off regulators without
sending a PON to eMMC is not a good idea.

Hence, allow upto 10sec for eMMC to be probed. If eMMC is
present it'd vote for the regulators and removing UFS vote
wouldn't cause it to suddenly turn off.

Change-Id: I54f915ef8708919ef1f0a85cc44de42b5df3710a
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>