treewide: Fix code errors detected using GCC 8.1.0
Use the latest version of GCC to take advantage of improved static analysis.
These issues appeared as warnings from the compiler.

Included in this commit is a buffer overflow fix (in touch_sw49408_prd.h),
as well as fixes for other lethal issues.

Compiler warnings should not be taken lightly.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>