msm: kgsl: adjust 81176a432d2b68e9f1f983a3b9eaae896ea4ae70 for msm8996
In that commit default_pwrlevel was removed, which might
not be a big issue for msm8998, but is an issue for 8996.

It has 8 pwrlevels and thus by setting initial gpu freq level
to 'num_pwrlevels - 1' gpu frequncy gets to... 27 MHz. Which
isn't very good for UX.

Adjust so it would pick 133 mhz instead.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>